moving the input component into a separate file for modularity
